pyproject.toml
[project]
name = "fastapi-pdf-qa"
version = "0.1.0"
description = ""
authors = [
    {name = "disha",email = "dishagopani05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "sentence-transformers (>=4.0.2,<5.0.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "fastapi_pdf_qa", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
